Delete original-arxiv_v6 index
This PR documents the change of the datavis-arxlive-copy domain on ES.
In particular, documents the removal of the original-arxiv_v6 index. An
associated snapshot is saved for reference. Also included is the
addition of the pipeline API, which is needed so we can use the arxiv_v6
index to reindex a test index. The pipeline API allows the removal of
fields while using the reindex file, which is needed so we can ignore
previously annotated data when testing.

closes #21