Change mapping to count word tokens for abstracts (#23)
This PR documents the need to change the `arxiv_v6` mapping so that the
token count for the `textBody_abstract_article` is computed on
ingestion. This meant we had to first reindex the `arxiv_v6` index to
another, temporary index, and then delete the old `arxiv_v6` index. Once
the reindexing had finished, we could reindex back to a new index named
`arxiv_v6` and delete the temporary index. A cumbersome and lengthy
process, but the only one that appears to work for computing analyzers
on existing indices.

For clarity, the process looks like:

reindex a -> b (with new mapping)
delete a
reindex b -> a (to rename)
delete b

closes #19