Force flag allows proceed when field exists
Supplying the force flag to arxlive_spotlight_annotation/src/bin/es/annotate.mjs
will now allow the script to continue, regardless of whether the
supplied field argument exists on the index's mapping or not.

This then allows the testAnnotationLocally npm script to name the
annotated field 'dbpedia_entities', which will mimic the arxiv_v6
schema so that the test index and arxiv_6 index can be queried using
the exact same request bodies.

closes #24

Also:
- cleanup arxlive_spotlight_annotation/src/bin/es/index.mjs (remove console.logs)
- cleanup arxlive_spotlight_annotation/src/node_modules/conf/mappings.mjs (remove redundant fields)
- cleanup arxlive_spotlight_annotation/src/node_modules/es/requests.mjs (remove stale imports)
- cleanup arxlive_spotlight_annotation/src/node_modules/es/snapshot.mjs (return request object instead of request body)