Add metadata upon ingestion and reindex domain
This PR introduces the ability to add metadata to the ES domain if the
user so desires. The metadata in question is documented in the issue
linked to this PR. The PR also documents the changes to the arxiv_v6
mappings and the different changes made to facilitate the data quality
aggregations currently underway. All changes to the mappings are
documented in elasticSearch/datavis-arxlive-copy/snapshot_v4/README.md

closes #47