Fix bulk update bug when no docs supplied
This PR fixes a bug wherein if the bulk update API is called and
no documents are supplied, the update fails as there is no requests
to be made.

Empty bulk updates may happen when we filter out documents which never
receives any annotations, due to garbage data or empty abstracts.

This PR also fixes a bug which occurs as a result of commander options
being treated as strings when they should have been parsed as integers.

fixes #66