Automate aggregation requests (#81)
PR to introduce code which automates the calling of the aggregation
requests. The script accepts ES related parameters (domain, index), and
paths indicating where to find the requests data and where to save the
responses. The inputs are the directories containing the requests, the
output path is a simple directory with each aggregation name saved as a
response json file.

closes #78