Refactor spec file to a script file
Refactors code from the spec file which generated data and instead
places this code in a script in the bin directory where it is better
suited. This PR also rearranges the data directory, as well as removes
the previously mentioned spec file as there were no actual tests being
run in it.

closes #119