Set up terraform state endpoint (#131)
PR to introduce the boilerplate for a fastify server and a single
endpoint, meant to test the status of the provisioned instances created
by the Terraform server.

- `down` (no resources provisioned)
- `scheduling` (resources have been provisioned using Terraform, but the
  instances aren't ready yet)
- `ready` (resources provisioned, return number of resources
  provisioned)

The PR also needs to set up status endpoints for the Terraform server,
seeing as this is where we can set status state depending on where in
the process order the execution is. The endpoints are the same, as
above, except that instead of `ready`, we use `up`, indicating that the
EC2 instances are provisioned and ready to recieve HTTP communication,
despite the fact the the Spotlight containers themselves might not be
ready to recieve input. I've decided to implement it this way as it
separates concerns - Terraform server is responsible for provision,
whereas Spotlight/Annotation server is responsible for annotation, and
therefore shouldn't be considered `ready` or `up` until the containers
are. Using `up` vs. `ready` helps distinguish these two states.

closes #130