Refactor logic using Xstate
Refactoring existing logic which uses `setInterval()` to determine
the state of provisioning and annotation and instead use Xstate -
results in cleaner code which is easier to test and maintain. Also
allows use to visualise the state of the machine.

The state machine will invoke processes which provision EC2 instances
and annotate ES endpoints. The processes wait until the spotlight server
finishes provisioning (promised based) before annotation, by polling
the status endpoint of the server every 15s.

Note: the tests for the Xstate machine are currently not working when
using mocha testing library. This is left as a TODO.

closes #138