Script for uploading CSV data to ES
Introduces a script which loads a CSV and creates documents for
each row in that CSV on an ElasticSearch index. Currently loads the
entire CSV to memory, so this approach is not possible for large CSV
files and future work is needed to implement this feature if the need
arises.

closes #142