Get count of docs from #indexToBucket
Get's the total number of documents that have been transferred to an
ES index from an S3 bucket by counting the docs used from the stream
iterator. Will give an exact count, as opposed to using the count API
from ElasticSearch, as this wasn't guaranteed to return the correct
result.

fixes #193