Change request type so send arrays of text (#190)
This PR changes the type of annotation request a spotlight worker
node recieves. Instead, the worker will recieve an array of text, which
means that the node itself will need to host not only a spotlight
container, but also a small server capable of handling the logic needed
for these requests. As a result, the worker nodes have much less
network overhead, and they're processing much larger chunks of data.

closes #189