Fix bug where docs weren't annotating
The bulk API for ElasticSearch wasn't waiting for the shards to refresh,
resulting in a significant portion of documents not being annotated at
all. By using the refresh=wait_for query paramater, we ensure that the
docs are both uploaded and ready to be annoated before launching the
service.

Also:
- make the "simple" S3 output format the default.

fixes #206